Added PlayerPawn class to project Added RotableActor class to project
PlayerPawn is a simple pawn that handle basic flying movements and rotations with WASD and mouse keys (need to configure inputs in project...).
PlayerPawn also holds a mouse manager, with a push on the CTRL key you can interact with RotableActors.

RotableActor define the basic comportment of an object that the player can rotate with the mouse.